check that that tp_pid exits, otherwise might inadvertedly kill engine
git-svn-id: https://vcs.maemo.org/svn/min/trunk@475 86ee150f-fb1f-4fea-b4f8-59f22ff7da1d

+++ min/src/tmc/tmc_msghnd.c
@@ -451,13 +451,15 @@ void gu_handle_message (TMC_t * tmc, const MsgBuffer * msg)
                 globaltcr.result_ = msg->param_;
                 STRCPY (globaltcr.desc_, msg->message_,
                         MaxTestResultDescription + 1);
-
-                retval = kill (tp_pid (&tmc->tpc_), SIGUSR2);
-                if (retval == -1)
-                        MIN_ERROR ("RET: Cannot kill child process");
-                else
-                        MIN_DEBUG ("RET: Killed child process");
-                break;
+		if (tp_pid (&tmc->tpc_)) {
+			retval = kill (tp_pid (&tmc->tpc_), SIGUSR2);
+			if (retval == -1)
+				MIN_ERROR ("RET: Cannot kill child process");
+			else
+				MIN_DEBUG ("RET: Killed child process");
+		} else
+			MIN_DEBUG ("RET: already killed child process");
+		break;
 
         case MSG_END:
                 gu_handle_end (tmc);
